Fixed amount and price of product datatype to float, added enum to choose for price weight  type

+++ AveCaesarApp/Models/ProductDishes.cs
@@ -14,6 +14,6 @@ namespace AveCaesarApp.Models
         public int DishId { get; set; }
         public Dish Dish { get; set; }
 
-        public int ProductAmount { get; set; }
+        public float ProductAmount { get; set; }
     }
 }
